requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
ipykernel==6.29.5
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
scikit-learn==1.5.2
setuptools==75.6.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
